move logic to npm:updateArtifacts
