+++ lib/modules/manager/npm/artifacts.ts
@@ -1,15 +1,21 @@
+import { isNonEmptyArray, isNonEmptyObject, isNumber } from '@sindresorhus/is';
 import upath from 'upath';
 import { logger } from '../../../logger';
 import { exec } from '../../../util/exec';
 import type { ExecOptions } from '../../../util/exec/types';
 import {
   ensureCacheDir,
+  getSiblingFileName,
+  localPathExists,
   readLocalFile,
   writeLocalFile,
 } from '../../../util/fs';
 import { regEx } from '../../../util/regex';
+import { matchRegexOrGlob } from '../../../util/string-match';
+import { parseSingleYaml } from '../../../util/yaml';
 import type { UpdateArtifact, UpdateArtifactsResult } from '../types';
 import { PNPM_CACHE_DIR, PNPM_STORE_DIR } from './constants';
+import type { PnpmWorkspaceFile } from './extract/types';
 import { getNodeToolConstraint } from './post-update/node-version';
 import { processHostRules } from './post-update/rules';
 import { lazyLoadPackageJson } from './post-update/utils';
@@ -25,13 +31,31 @@ const versionWithHashRegString = '^(?<version>.*)\\+(?<hash>.*)';
 // Execute 'corepack use' command for npm manager updates
 // This step is necessary because Corepack recommends attaching a hash after the version
 // The hash is generated only after running 'corepack use' and cannot be fetched from the npm registry
-export async function updateArtifacts({
-  packageFileName,
-  config,
-  updatedDeps,
-  newPackageFileContent: existingPackageFileContent,
-}: UpdateArtifact): Promise<UpdateArtifactsResult[] | null> {
-  logger.debug(`npm.updateArtifacts(${packageFileName})`);
+export async function updateArtifacts(
+  updateArtifactsConfig: UpdateArtifact,
+): Promise<UpdateArtifactsResult[] | null> {
+  logger.debug(`npm.updateArtifacts(${updateArtifactsConfig.packageFileName})`);
+  let res: UpdateArtifactsResult[] = [];
+  res.push((await handlePackageManagerUpdates(updateArtifactsConfig)) ?? {});
+  res.push((await updatePnpmWorkspace(updateArtifactsConfig)) ?? {});
+
+  res = res.filter(isNonEmptyObject);
+  if (res.length === 0) {
+    return null;
+  }
+
+  return res;
+}
+
+async function handlePackageManagerUpdates(
+  updateArtifactsConfig: UpdateArtifact,
+): Promise<UpdateArtifactsResult | null> {
+  const {
+    packageFileName,
+    config,
+    updatedDeps,
+    newPackageFileContent: existingPackageFileContent,
+  } = updateArtifactsConfig;
   const packageManagerUpdate = updatedDeps.find(
     (dep) => dep.depType === 'packageManager',
   );
@@ -101,25 +125,141 @@ export async function updateArtifacts({
       return null;
     }
     logger.debug('Returning updated package.json');
-    return [
-      {
-        file: {
-          type: 'addition',
-          path: packageFileName,
-          contents: newPackageFileContent,
-        },
+    return {
+      file: {
+        type: 'addition',
+        path: packageFileName,
+        contents: newPackageFileContent,
       },
-    ];
+    };
   } catch (err) {
     logger.warn({ err }, 'Error updating package.json');
     await resetNpmrcContent(pkgFileDir, npmrcContent);
-    return [
-      {
-        artifactError: {
-          fileName: packageFileName,
-          stderr: err.message,
-        },
+    return {
+      artifactError: {
+        fileName: packageFileName,
+        stderr: err.message,
       },
-    ];
+    };
   }
+}
+
+/**
+ * Update the minimumReleaseAgeExclude setting in pnpm-workspace.yaml if needed
+ */
+async function updatePnpmWorkspace(
+  updateArtifactsConfig: UpdateArtifact,
+): Promise<UpdateArtifactsResult | null> {
+  const pnpmShrinkwrap = updateArtifactsConfig.updatedDeps[0].managerData
+    ?.pnpmShrinkwrap as string;
+  const lockFileDir = upath.dirname(pnpmShrinkwrap);
+  const lockFileName = upath.join(lockFileDir, 'pnpm-lock.yaml');
+  const pnpmWorkspaceFilePath = getSiblingFileName(
+    lockFileName,
+    'pnpm-workspace.yaml',
+  );
+
+  if (!(await localPathExists(pnpmWorkspaceFilePath))) {
+    return null;
+  }
+
+  const oldContent = (await readLocalFile(pnpmWorkspaceFilePath, 'utf8'))!;
+  let packageFileContent = oldContent;
+  let updated = false;
+
+  for (const upgrade of updateArtifactsConfig.updatedDeps) {
+    // if (!upgrade.isVulnerabilityAlert) {
+    //   continue;
+    // }
+    logger.debug('updatePnpmWorkspace()');
+
+    const pnpmWorkspace =
+      parseSingleYaml<PnpmWorkspaceFile>(packageFileContent);
+    if (!isNumber(pnpmWorkspace?.minimumReleaseAge)) {
+      continue;
+    }
+
+    if (!isNonEmptyArray(pnpmWorkspace.minimumReleaseAgeExclude)) {
+      logger.debug('Adding new exclude block');
+      // add minimumReleaseAgeExclude
+      const addedStr = `
+        minimumReleaseAgeExclude:
+         - ${upgrade.depName}@${upgrade.newValue}
+      `;
+      const newContent = oldContent + addedStr;
+      await writeLocalFile(pnpmWorkspaceFilePath, newContent);
+      packageFileContent = newContent;
+      updated = true;
+      continue;
+    }
+    // check if exlcude setting exists for the dep
+    let matchingSetting: {
+      match: boolean;
+      matchType?: 'pattern' | 'all-versions' | 'single-versions';
+      settingStr: string;
+    } = { match: false, settingStr: '' };
+    for (const setting of pnpmWorkspace.minimumReleaseAgeExclude ?? []) {
+      const matchingRes = checkExcludeSetting(setting, upgrade.depName!);
+      if (matchingRes.match) {
+        matchingSetting = {
+          match: matchingRes.match,
+          matchType: matchingRes.matchType,
+          settingStr: setting,
+        };
+      }
+    }
+
+    if (matchingSetting.match) {
+      logger.debug('Matching setting found');
+      if (matchingSetting.matchType === 'single-versions') {
+        logger.debug('Matching setting found, appending ||');
+        // need to find and replace the old setting by appending || <newValue>
+        const newSetting =
+          matchingSetting.settingStr + ` || ${upgrade.newValue}`;
+        const newContent = oldContent.replace(
+          matchingSetting.settingStr,
+          newSetting,
+        );
+        await writeLocalFile(pnpmWorkspaceFilePath, newContent);
+        packageFileContent = newContent;
+        updated = true;
+      }
+    }
+  }
+
+  if (!updated) {
+    return null;
+  }
+
+  return {
+    file: {
+      type: 'addition',
+      path: pnpmWorkspaceFilePath,
+      contents: packageFileContent,
+    },
+  };
+}
+
+function checkExcludeSetting(
+  setting: string,
+  depName: string,
+): {
+  match: boolean;
+  matchType?: 'pattern' | 'all-versions' | 'single-versions';
+} {
+  if (setting.includes(depName)) {
+    if (!setting.includes('@')) {
+      return { match: true, matchType: 'all-versions' };
+    }
+    return { match: true, matchType: 'single-versions' };
+  }
+
+  // check if setting is a pattern
+  // TODO: use getRegexOrGlobPredicate method
+  const res = matchRegexOrGlob(depName, setting);
+  if (res) {
+    return { match: true, matchType: 'pattern' };
+  }
+
+  return { match: false };
 }
