+++ lib/modules/manager/npm/update/dependency/index.ts
@@ -12,7 +12,7 @@ import type {
 } from '../../extract/types';
 import type { NpmDepType, NpmManagerData } from '../../types';
 import { getNewGitValue, getNewNpmAliasValue } from './common';
-import { updatePnpmCatalogDependency, updatePnpmWorkspace } from './pnpm';
+import { updatePnpmCatalogDependency } from './pnpm';
 import { updateYarnrcCatalogDependency } from './yarn';
 
 function renameObjKey(
@@ -114,21 +114,16 @@ function replaceAsString(
   throw new Error();
 }
 
-export async function updateDependency({
+export function updateDependency({
   fileContent,
   upgrade,
-}: UpdateDependencyConfig): Promise<string | null> {
+}: UpdateDependencyConfig): string | null {
   if (upgrade.depType?.startsWith('pnpm.catalog')) {
     return updatePnpmCatalogDependency({ fileContent, upgrade });
   }
   if (upgrade.depType?.startsWith('yarn.catalog')) {
     return updateYarnrcCatalogDependency({ fileContent, upgrade });
   }
-
-  // check if it's security upgrade & pnpm-workspace has minimumReleaseAge set
-  // if (upgrade.isVulnerabilityAlert) {
-  await updatePnpmWorkspace({ fileContent, upgrade });
-  // }
 
   const { depType, managerData } = upgrade;
   const depName: string = managerData?.key ?? upgrade.depName;
